Add dedicated monitor number assignments
Either Windows or AutoHotkey sometimes decides to change how monitor
numbers are assigned. When this happens, the user has to update the script
with the new monitor numbers so that all of the window management features
work as expected. This was a pain in the butt.

Now, when this happens, the user will only have to update the values of
the new dedicated monitor number assignment variables. This is slightly
less of a pain in the butt and is still a better solution than hard-coding
the monitor coordinates manually.